--- encryption-service/impl/authn/authstore_mock.go
@@ -1,33 +1,0 @@
-// Copyright 2021 CYBERCRYPT
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-package authn
-
-import (
-	"context"
-
-	"encryption-service/interfaces"
-)
-
-type AuthStoreMock struct {
-	NewTransactionFunc func(ctx context.Context) (authStoreTx interfaces.AuthStoreTxInterface, err error)
-	CloseFunc          func()
-}
-
-func (as *AuthStoreMock) NewTransaction(ctx context.Context) (interfaces.AuthStoreTxInterface, error) {
-	return as.NewTransactionFunc(ctx)
-}
-
-func (as *AuthStoreMock) Close() {
-	as.CloseFunc()
-}
